+++ packages/react/src/display/display.tsx
@@ -6,6 +6,25 @@ import { ResponsiveType } from '../shared';
 export const displaySizes = ['s', 'm', 'l', 'xl'] as const;
 export type DisplaySizesType = typeof displaySizes[number];
 
+export const displayTagNames = [
+  'blockquote',
+  'dd',
+  'div',
+  'dt',
+  'h1',
+  'h2',
+  'h3',
+  'h4',
+  'h5',
+  'h6',
+  'li',
+  'p',
+  'span',
+  'td',
+  'th',
+] as const;
+export type DisplayTagNamesType = typeof displayTagNames[number];
+
 export const displayVariants = ['primary', 'secondary'] as const;
 export type DisplayVariantsType = typeof displayVariants[number];
 
@@ -14,6 +33,7 @@ export interface DisplayProps {
   size?: ResponsiveType<DisplaySizesType>;
   light?: boolean;
   className?: string;
+  tagName?: DisplayTagNamesType;
   variant?: DisplayVariantsType;
 }
 
@@ -55,9 +75,11 @@ const Display: React.FC<DisplayProps> = ({
   size = 'xl',
   light = false,
   className = '',
+  tagName = 'span',
   variant = 'primary',
   ...otherProps
 }) => {
+  const Component = tagName;
   const sizes = typeof size === 'object' ? size : { mobile: size, tablet: size, desktop: size };
   const fontSizes = {
     mobile: textStyles[sizes.mobile].mobile,
@@ -66,7 +88,7 @@ const Display: React.FC<DisplayProps> = ({
   };
 
   return (
-    <span
+    <Component
       className={clsx(
         'db-display',
         responsiveClassNames(fontSizes, 'db-'),
@@ -77,7 +99,7 @@ const Display: React.FC<DisplayProps> = ({
       {...otherProps}
     >
       {children}
-    </span>
+    </Component>
   );
 };